move addToCartButton into its own component
